fix flags
