+++ internal/storage/metrics/storage.go
@@ -12,20 +12,11 @@ import (
 // tmpFile is the temporary file to store the metrics
 const tmpFile = "/tmp/sysmon.json"
 
-// Storage defines the interface for storing the metrics of the system
-type Storage interface {
-	// Get returns the metrics of the system from the storage
-	Get(ctx context.Context) (models.Metrics, error)
-
-	// Set stores the metrics of the system
-	Set(ctx context.Context, m models.Metrics) error
-}
-
 // storage implements the Storage interface
 type storage struct{}
 
 // NewStorage returns a new instance of Storage
-func NewStorage() Storage {
+func NewStorage() *storage {
 	return &storage{}
 }
 
@@ -45,6 +36,7 @@ func (s *storage) Get(_ context.Context) (models.Metrics, error) {
 	return m, nil
 }
 
+// Set stores the metrics of the system
 func (s *storage) Set(_ context.Context, m models.Metrics) error {
 	str, err := json.Marshal(m)
 	if err != nil {
